Add drill planner job
Add a MVP as the first step to handle drill planning in everest. The
OR-tool will be used in conjuction with a greedy algorithm, for any case
that the OR-tool can not find a solution within the max solve time.
Multiple issues have been created that will change how the model is
structured in order to optimize the layout.

Co-authors:
 - Eivind Jahren
 - Lars Petter Øren Hauge