Add rate/phase/duration constraints job
Add a job to Everest that will handle well operations, including rate,
phase and duration sensitivities. The user must spesify bounds for
all controls, and has the option to input constant values for well
events. As a first pass the user must input phase/rate and duration
for each event, but this should probably change in the future, however
more complex validation is required to make that possible.