Resolve priorities
Only related to drill planner:

The solves should not have a hard constraint on the priorities, they
should try to find the most optimal schedule. And in a case where a
lower priority well can be drilled before a higher priority well without
affecting it, the lower priority well will be drilled first. However, we
wish to honour the priority from the optimizer, and thus we shift the
completion date of the lower priority well to match the higher priority
well.