Create an abstraction for schedule verification
We utilize a class (FieldManager) in the drillplanner to create an
abstraction layer between the configuration and the optimizers. The
FieldManager contains all the constraints from the configuration and
answers if the provided schedule is valid given the constraints.

This substitues the verify_constraints function as well.

The FieldManager will be more utilized in the next implementation of the
or-tool solver