Rework or-tools implementation in drillplanner
The commit includes a complete rework of how the model is created, with
the main change of using OptionalIntervalVar instead of IntervalVar for
each task created - and the benefits that comes with it.

With IntervalVar's each well and rig pair was given a task each, and in
any case the first task for a well completed was the one chosen. This
meant that every well,rig task had to be completed within the horizon,
even though only the first was of any interest. Furthermore the handling
of slot unavailability was quite poor and difficult to understand.

The other major change is the removal of a hard constraint on well
priorities.

Using OptionalIntervalVar's we specify one task per well,slot,rig
combination and the `presence` variable is used to handle both number of
tasks per well (1) and number of tasks per slot (<=1).

Slot and rig availability is handled in the same manner, with
NoOverlap definitions added to the tasks.

With the removal of the hard constraint of well priorities, the modified
objective function behave a bit differently, allowing a lower priority
well to be drilled prior to a higher priority well given that the higher
priority well is not affected.