Change FieldSchedule datastructure
We simplify the data contained in a FieldSchedule, and leave all
validations up the FieldManager class. Currently there's a bit switching
between the data type (either strings or instances of a Rig, Slot or
Well class), and we want to keep those contained within the FieldManager
class.

The changes also simplifies the writing of tests, and eventually should
releave us of all utility functions such as create_schedule_events.