Check outputfile writable
We first check that the filepath given from the user is valid during
argparse parsing. Then we do another check once the job is completed as
things may have changed in the meantime.

We could do a try-except on 'open with' for testing of the file during
parsing, but this also creates an empty file which a user might try to
access, inherently making it unwritable once the job has completed.