Add date to dynamic perforations only
ResInsight writes LAS files for each simulation date when the property was
extracted. When using dynamic and static properties this results in at least
two LAS files. In this PR we read all the LAS files associated with the current
well, we assume these LAS files have the same length and concatenate these arrays.

The existing test is extended by covering this use case (using both dynamic and static).

Add message for deprecated use of date in config