Add date to dynamic perforations only
ResInsight writes LAS files for each simulation date when the property was
extracted. When using dynamic and static properties this results in at least
two LAS files. In this PR we read all the LAS files associated with the current
well and join these arrays/intervals together based on common DEPTH  TVDMSL  TVDRKB.
NOTE: Inner join is OK since non-existing intervals will be filtered out later on
based on min-max of the properties (in case outer join is used with default value
-999.99 since this lies (most likely) outside the min-max range of the property);
the key thing is that we join on the correct depth values.

The existing test is extended by covering this use case (using both dynamic and static).