Only install test dependencies in testkomodo
Running `pip install .[test]` also runs pip install
on normal dependencies. When we have specified a https-address
for fmu-steaclient in pyproject.toml, it will prefer that over
any fmu-steaclient already existing in its venv, which is wrong
when doing komodo integration tests.

By extracting the test dependencies only, and on-the-fly, we
can install only these and ensure that fmu-steaclient is not
attempted upgraded.